+++ ExperienceInformation/ExperienceInformation.Web/UI_EnergyInnovation/OperationEnergySavingSuggestions.aspx.cs
@@ -20,8 +20,21 @@ namespace ExperienceInformation.Web.UI_EnergyInnovation
             base.InitComponts();
             if (!IsPostBack)
             {
-
+#if DEBUG
+                List<string> m_DataValidIdItems = new List<string>() { "zc_nxjc_qtx_efc", "zc_nxjc_byc_byf" };
+                AddDataValidIdGroup("ProductionOrganization", m_DataValidIdItems);
+                mPageOpPermission = "0000";
+#endif
             }
+        }
+        /// <summary>
+        /// 增删改查权限控制
+        /// </summary>
+        /// <returns></returns>
+        [WebMethod]
+        public static char[] AuthorityControl()
+        {
+            return mPageOpPermission.ToArray();
         }
         [WebMethod]
         public static string GetEnergySavingSuggestionsInfo(string myCreateYear, string myKeyword)
@@ -47,45 +60,66 @@ namespace ExperienceInformation.Web.UI_EnergyInnovation
         [WebMethod]
         public static string AddEnergySavingSuggestions(string mySuggestionsName, string myKeyword, string myPostName, string mySuggestionsText, string myPropounder, string myProposedTime, string myRemarks)
         {
-            if (mUserId != "")
+            if (mPageOpPermission.ToArray()[1] == '1')
             {
-                int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.AddEnergySavingSuggestions(mySuggestionsName, myKeyword, myPostName, "", SuggestionsType, EnergySavingSuggestions,
-                       mySuggestionsText, myPropounder, myProposedTime, mUserId, myRemarks);
-                int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
-                return m_Result.ToString();
+                if (mUserId != "")
+                {
+                    int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.AddEnergySavingSuggestions(mySuggestionsName, myKeyword, myPostName, "", SuggestionsType, EnergySavingSuggestions,
+                           mySuggestionsText, myPropounder, myProposedTime, mUserId, myRemarks);
+                    int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
+                    return m_Result.ToString();
+                }
+                else
+                {
+                    return "非法的用户操作!";
+                }
             }
             else
             {
-                return "非法的用户操作!";
+                return "该用户没有添加权限！";
             }
         }
         [WebMethod]
         public static string ModifyEnergySavingSuggestions(string mySuggestionsId, string mySuggestionsName, string myKeyword, string myPostName, string mySuggestionsText, string myPropounder, string myProposedTime, string myRemarks)
         {
-            if (mUserId != "")
+            if (mPageOpPermission.ToArray()[2] == '1')
             {
-                int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.ModifyEnergySavingSuggestionsById(mySuggestionsId, mySuggestionsName, myKeyword, myPostName, "", SuggestionsType, EnergySavingSuggestions,
-                       mySuggestionsText, myPropounder, myProposedTime, mUserId, myRemarks);
-                int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
-                return m_Result.ToString();
+                if (mUserId != "")
+                {
+                    int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.ModifyEnergySavingSuggestionsById(mySuggestionsId, mySuggestionsName, myKeyword, myPostName, "", SuggestionsType, EnergySavingSuggestions,
+                           mySuggestionsText, myPropounder, myProposedTime, mUserId, myRemarks);
+                    int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
+                    return m_Result.ToString();
+                }
+                else
+                {
+                    return "非法的用户操作!";
+                }
             }
             else
             {
-                return "非法的用户操作!";
+                return "该用户没有修改权限！";
             }
         }
         [WebMethod]
         public static string DeleteEnergySavingSuggestionsById(string mySuggestionsId)
         {
-            if (mUserId != "")
+            if (mPageOpPermission.ToArray()[3] == '1')
             {
-                int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.DeleteEnergySavingSuggestionsById(mySuggestionsId);
-                int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
-                return m_Result.ToString();
+                if (mUserId != "")
+                {
+                    int m_EnergySavingSuggestionsText = ExperienceInformation.Service.EnergyInnovation.EnergySavingSuggestions.DeleteEnergySavingSuggestionsById(mySuggestionsId);
+                    int m_Result = m_EnergySavingSuggestionsText > 0 ? 1 : m_EnergySavingSuggestionsText;
+                    return m_Result.ToString();
+                }
+                else
+                {
+                    return "非法的用户操作!";
+                }
             }
             else
             {
-                return "非法的用户操作!";
+                return "该用户没有删除权限！";
             }
         }
     }
